Spring: Implement view pad notes
This view shows a pageable list of notes of a pad. The paging menu is
not in the view yet. Also this view needs a pad filter. Showing all
notes without a pad is not implemented yet. The create note button does
not yet preselect the respective pad.

See also: komarserjio/notejam#52